api/requirements.txt
flask
pydash
python-dotenv
flask_cors
flask-praetorian
pendulum
Flask-SQLAlchemy
